Move UiMessage(Manager)
