+++ ui/episode/details/src/main/java/app/tivi/episodedetails/EpisodeDetailsPresenter.kt
@@ -8,8 +8,8 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
-import app.tivi.api.UiMessage
-import app.tivi.api.UiMessageManager
+import app.tivi.common.compose.UiMessage
+import app.tivi.common.compose.UiMessageManager
 import app.tivi.common.compose.rememberCoroutineScope
 import app.tivi.domain.interactors.RemoveEpisodeWatch
 import app.tivi.domain.interactors.RemoveEpisodeWatches
