+++ ui/episode/track/src/main/java/app/tivi/episode/track/EpisodeTrackPresenter.kt
@@ -11,8 +11,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import app.tivi.api.UiMessage
-import app.tivi.api.UiMessageManager
+import app.tivi.common.compose.UiMessage
+import app.tivi.common.compose.UiMessageManager
 import app.tivi.common.compose.rememberCoroutineScope
 import app.tivi.domain.interactors.AddEpisodeWatch
 import app.tivi.domain.interactors.UpdateEpisodeDetails
