+++ ui/library/src/main/java/app/tivi/home/library/LibraryUiState.kt
@@ -5,7 +5,7 @@ package app.tivi.home.library
 
 import androidx.compose.runtime.Stable
 import app.cash.paging.compose.LazyPagingItems
-import app.tivi.api.UiMessage
+import app.tivi.common.compose.UiMessage
 import app.tivi.data.compoundmodels.LibraryShow
 import app.tivi.data.models.SortOption
 import app.tivi.data.models.TraktUser
