+++ ui/search/src/main/java/app/tivi/home/search/SearchPresenter.kt
@@ -10,8 +10,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import app.tivi.api.UiMessage
-import app.tivi.api.UiMessageManager
+import app.tivi.common.compose.UiMessage
+import app.tivi.common.compose.UiMessageManager
 import app.tivi.common.compose.rememberCoroutineScope
 import app.tivi.data.models.TiviShow
 import app.tivi.domain.interactors.SearchShows
