+++ ui/upnext/src/main/java/app/tivi/home/upnext/UpNextUiState.kt
@@ -5,7 +5,7 @@ package app.tivi.home.upnext
 
 import androidx.compose.runtime.Stable
 import app.cash.paging.compose.LazyPagingItems
-import app.tivi.api.UiMessage
+import app.tivi.common.compose.UiMessage
 import app.tivi.data.compoundmodels.UpNextEntry
 import app.tivi.data.models.SortOption
 import app.tivi.data.models.TraktUser
